Moved Controlsystem functions to controlsystem class fro controller.cpp + changed loop from double to unsigned long so int the same all the way through + minor formatting changes

+++ Arduino/Dynamics.h
@@ -7,7 +7,7 @@
 class Dynamics
 {
 public:
-	void UpdateDynamics(double& Looptime);
+	void UpdateDynamics(unsigned long& Looptime);
 
 private:
 	double _prevVel1{0}, _prevVel2{0}, _prevVel3{0};
